requirements.txt
streamlit
pandas
yfinance
tefas-crawler
plotly
sqlalchemy
openpyxl
streamlit-option-menu
